--- examples/src/test/resources/application.conf
@@ -1,21 +1,0 @@
-akka {
-  loggers = ["akka.event.slf4j.Slf4jLogger"]
-  loglevel = "DEBUG"
-  logging-filter = "akka.event.slf4j.Slf4jLoggingFilter"
-  logger-startup-timeout = 15s
-
-  actor {
-    debug {
-      lifecycle = off
-      receive = off
-    }
-  }
-
-  test {
-    single-expect-default = 20s
-  }
-
-  kafka.consumer {
-    stop-timeout = 3 s
-  }
-}
